update schema version to 1.6 added type change for UTMEasting & UTMNorthing to SchemaUpdateService.
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@9241 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/datamodel/LocalityDetail.java
@@ -83,8 +83,8 @@ public class LocalityDetail extends DataModelObjBase
 
     
     // UTM Fields
-    protected Float                 utmEasting;
-    protected Float                 utmNorthing;
+    protected BigDecimal            utmEasting;
+    protected BigDecimal            utmNorthing;
     protected Integer               utmFalseEasting;
     protected Integer               utmFalseNorthing;
     protected String                utmDatum;
@@ -570,7 +570,7 @@ public class LocalityDetail extends DataModelObjBase
      * @return the utmEasting
      */
     @Column(name = "UtmEasting", unique = false, nullable = true, updatable = true, insertable = true)
-    public Float getUtmEasting()
+    public BigDecimal getUtmEasting()
     {
         return utmEasting;
     }
@@ -578,7 +578,7 @@ public class LocalityDetail extends DataModelObjBase
     /**
      * @param utmEasting the utmEasting to set
      */
-    public void setUtmEasting(Float utmEasting)
+    public void setUtmEasting(BigDecimal utmEasting)
     {
         this.utmEasting = utmEasting;
     }
@@ -587,7 +587,7 @@ public class LocalityDetail extends DataModelObjBase
      * @return the utmNorthing
      */
     @Column(name = "UtmNorthing", unique = false, nullable = true, updatable = true, insertable = true)
-    public Float getUtmNorthing()
+    public BigDecimal getUtmNorthing()
     {
         return utmNorthing;
     }
@@ -595,7 +595,7 @@ public class LocalityDetail extends DataModelObjBase
     /**
      * @param utmNorthing the utmNorthing to set
      */
-    public void setUtmNorthing(Float utmNorthing)
+    public void setUtmNorthing(BigDecimal utmNorthing)
     {
         this.utmNorthing = utmNorthing;
     }
